# Implement full pixel triplets