--- a/HeterogeneousCore/CUDAUtilities/interface/GPUVecArray.h
+++ b/HeterogeneousCore/CUDAUtilities/interface/GPUVecArray.h
@@ -87,7 +87,7 @@ namespace GPU {
     inline constexpr T &operator[](int i) { return m_data[i]; }
     inline constexpr const T &operator[](int i) const { return m_data[i]; }
     inline constexpr void reset() { m_size = 0; }
-    inline constexpr int capacity() const { return maxSize; }
+    inline static constexpr int capacity() { return maxSize; }
     inline constexpr T const *data() const { return m_data; }
     inline constexpr void resize(int size) { m_size = size; }
     inline constexpr bool empty() const { return 0 == m_size; }