--- a/RecoLocalTracker/SiPixelClusterizer/plugins/gpuClustering.h
+++ b/RecoLocalTracker/SiPixelClusterizer/plugins/gpuClustering.h
@@ -12,6 +12,10 @@
 
 namespace gpuClustering {
 
+#ifdef GPU_DEBUG
+  __device__ uint32_t gMaxHit=0;
+#endif
+
   __global__ void countModules(uint16_t const* __restrict__ id,
                                uint32_t* __restrict__ moduleStart,
                                int32_t* __restrict__ clusterId,
@@ -272,8 +276,13 @@ namespace gpuClustering {
       nClustersInModule[thisModuleId] = foundClusters;
       moduleId[blockIdx.x] = thisModuleId;
 #ifdef GPU_DEBUG
+      if (foundClusters>gMaxHit) {
+         gMaxHit = foundClusters;
+         if (foundClusters>8) printf("max hit %d in %d\n",foundClusters, thisModuleId);
+      }
+#endif
+#ifdef GPU_DEBUG
       if (thisModuleId % 100 == 1)
-        if (threadIdx.x == 0)
           printf("%d clusters in module %d\n", foundClusters, thisModuleId);
 #endif
     }