--- a/RecoLocalTracker/SiPixelRecHits/plugins/gpuPixelRecHits.h
+++ b/RecoLocalTracker/SiPixelRecHits/plugins/gpuPixelRecHits.h
@@ -25,10 +25,32 @@ namespace gpuPixelRecHits {
     // The whole gimnastic here of copying or not is a pure heuristic exercise that seems to produce the fastest code with the above signature
     // not using views (passing a gazzilion of array pointers) seems to produce the fastest code (but it is harder to mantain)  
 
+    assert(phits);
+    assert(cpeParams);
+
     auto& hits = *phits;
 
     auto const digis = *pdigis; // the copy is intentional!
     auto const & clusters = *pclusters;
+
+    // copy average geometry corrected by beamspot . FIXME (move it somewhere else???)
+    if (0==blockIdx.x) {
+      auto & agc = hits.averageGeometry();
+      auto const & ag  = cpeParams->averageGeometry();
+      for(int il=threadIdx.x, nl=TrackingRecHit2DSOAView::AverageGeometry::numberOfLaddersInBarrel; il<nl; il+=blockDim.x) {
+        agc.ladderZ[il] = ag.ladderZ[il] - bs->z; 
+        agc.ladderX[il] = ag.ladderX[il] - bs->x;
+        agc.ladderY[il] = ag.ladderY[il] - bs->y;
+        agc.ladderR[il] = sqrt(agc.ladderX[il]*agc.ladderX[il] + agc.ladderY[il]*agc.ladderY[il] );
+        agc.ladderMinZ[il] = ag.ladderMinZ[il] - bs->z;
+        agc.ladderMaxZ[il] = ag.ladderMaxZ[il] - bs->z;
+      }
+      if(0==threadIdx.x) {
+         agc.endCapZ[0] = ag.endCapZ[0] - bs->z;
+         agc.endCapZ[1] = ag.endCapZ[1] - bs->z;
+//         printf("endcapZ %f %f\n",agc.endCapZ[0],agc.endCapZ[1]);
+      }
+    }
 
     // to be moved in common namespace...
     constexpr uint16_t InvId = 9999;  // must be > MaxNumModules