… RecoPixelVertexing/PixelTriplets/plugins/BuildFile.xml
… RecoPixelVertexing/PixelTriplets/plugins/CAHitQuadrupletGeneratorKernels.h
… RecoPixelVertexing/PixelTriplets/plugins/gpuPixelDoublets.h
… RecoPixelVertexing/PixelTriplets/plugins/gpuPixelDoubletsAlgos.h