Implement "full" pixel triplets reconstruction

default is still quadruplets: no regression w/r/t baseline. eventually a bit slower due to additional infrastructure

to switch triplets: set
```process.pixelTracksHitQuadruplets.minHitsPerNtuplet = 3```


new configurables
```
# set to False to add jumping doublets in the forward (just FPIX1,FPIX3)
process.pixelTracksHitQuadruplets.noForwardTriplets = True
# set to True to enable three iterations of _find_ntuples_ 
# each NOT using doublets already used in previous ones
process.pixelTracksHitQuadruplets.doIterations = False
```

MTV for Single Muon pt-flat ideal
http://innocent.home.cern.ch/innocent/RelVal/SingleMuFlatIdeal_NF/plots_pixel_pixel/effandfakePtEtaPhi.pdf

MTV for Single Muon pt50 realistic 
http://innocent.home.cern.ch/innocent/RelVal/SingleMu50Real_gpuNF/plots_pixel_pixel/effandfakePtEtaPhi.pdf
MTV for ttbar PU50 ideal
http://innocent.home.cern.ch/innocent/RelVal/pixOnlyPU50_gpuNF/plots_pixel_pixel/effandfakePtEtaPhi.pdf

timing (usual HLTPhysics 2018) after f31f13c
on workergpu11 Tesla V100-PCIE-16GB
baseline  1406.5 ±  12.0 ev/s
quadruplets 1261.6 ±   5.0 ev/s
triplets  1056.6 ±   1.7 ev/s
forward-triplets  1012.6 ±   3.8 ev/s
iterative forward-triplets 859.2 ±   1.9 ev/s

NB this is a V100: the cost of fitting is lower: on T4 I would expect a worse performance for adding forward-triplets as it mostly add triplets to be fit.
Confirmed my hypothesis that iterations will make the workflow just slower.
I will try to make optional some of the additional operations needed by iterations to gain speed

I would consider this as a reference point.
Optimization can come later