@VinInn 2019-06-26
reduced the phi cut for the jumping doublets: equivalent to a pt cut larger than 1GeV
for reference this is how looks like 1GeV muon realistic (2018)
http://innocent.home.cern.ch/innocent/RelVal/SingleMu1_ftriplRC/plots_pixel_pixel/effandfakePtEtaPhi.pdf
orange is with the old larger cuts

ttbar ideal is barely affected below 1GeV (with a strange increase in fake rate to-be-understood)
http://innocent.home.cern.ch/innocent/RelVal/pixOnlyPU50_gpuNFOpt/plots_pixel_pixel/effandfakePtEtaPhi.pdf
here stricter cuts is black