@VinInn 2019-06-27
quadruplets are now in the 1350 Hz
triplets still just above 1KHz (before were affected by overflows in data)