@VinInn 2019-06-28
Once I reduced the total number of running kernels the throughput for Quadruplets is the same as the baseline. Triplets ~1.1KHz.

Now we need to understand how to deliver this stuff to downstream (mostly CPU seeds)