@fwyzard 2019-07-02
:construction: Validation running at workergpu05:/mnt/home/fwyzard/validation/run.PYjPOpu5Sg ...