@VinInn 2019-07-05
counters for 5000 events of HLTPhysics

||Type | nEvents | nHits | nCells | nTuples | nFitTacks  |  nGoodTracks | nUsedHits | nDupHits | nKilledCells | nEmptyCells | nZeroTrackCells ||
||FTrip; |  5000 |  16482.7|  187577.5|  7421.4|  6179.8|  1158.3|  7298.1|  3309.4|  861.9|  0.933|  0.940||
||Quad |  5000 ||  16482.7|  94957.2|  698.6|  566.7|  345.4|  1414.1|  4.6|  333.2|  0.945|  0.986||
||Tripl |  5000 ||  16482.7|  160469.0|  5888.3|  4902.6|  995.3|  6199.9|  2780.9|  357.3|  0.938|  0.942||