@fwyzard 2019-07-10
:construction: Validation running at workergpu08:/mnt/home/fwyzard/validation/run.vJpHv7rJKN ...