@fwyzard 2019-07-31
I have run over the usual realistic and design TTbar sample with multiple options:
  - default configuration
  - enable triplets:
    ```python
    # switch on triplets
    process.pixelTracksHitQuadruplets.minHitsPerNtuplet = 3
    ```
  - enable triplets and iterations: same as above, plus
    ```
    # (only for triplets) enable three iterations of `kernel_find_ntuplets`
    #   - starting from different layers
    #   - skipping doublets already used in previous ones
    process.pixelTracksHitQuadruplets.doIterations = True
    ```
  - default configuration, with "jumping pairs":
    ```
    # enable "jumping doublets" in the forward (just FPIX1, FPIX3)
    process.pixelTracksHitQuadruplets.includeJumpingForwardDoublets = True
    ```
  - enable triplets, with "jumping pairs" (see above)
  - enable triplets and iterations, with "jumping pairs" (see above)

In each case I did not spot any difference enabling the iterations.
Is it expected ?