@fwyzard 2019-07-31
## design, no "jumping pairs"
![design-default_trackEffandfakePtEtaPhi](https://user-images.githubusercontent.com/4069793/62216245-ac69c000-b3a8-11e9-8576-d59392422d88.png)

## design, with "jumping pairs"
![design-jumping_pairs_trackEffandfakePtEtaPhi](https://user-images.githubusercontent.com/4069793/62216263-b4296480-b3a8-11e9-9386-7ca548f2dd3d.png)

## realistic, no "jumping pairs"
![realistic-default_trackEffandfakePtEtaPhi](https://user-images.githubusercontent.com/4069793/62216292-c1deea00-b3a8-11e9-9ff8-235b9f9f82f6.png)

## realistic, with "jumping pairs"
![realistic-jumping_pairs_trackEffandfakePtEtaPhi](https://user-images.githubusercontent.com/4069793/62216303-c6a39e00-b3a8-11e9-9bf1-f3df5dd86f07.png)