@fwyzard 2019-08-02
## Throughput on hilton with a T4
```
2 CPUs:
  0: Intel(R) Xeon(R) Gold 6130 CPU @ 2.10GHz (16 cores, 32 threads)
  1: Intel(R) Xeon(R) Gold 6130 CPU @ 2.10GHz (16 cores, 32 threads)

1 visible NVIDIA GPUs:
  0: Tesla T4 (UUID: GPU-0e44f21a-ab43-7bf2-08d7-8cf1c32dd05b)
```

Measured on the usual pileup 50 L1-accepted data.

### Before (reference)
```
Running 4 times over 10200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
   983.0 ±   0.3 ev/s (10000 events)
   987.0 ±   0.3 ev/s (10000 events)
   975.3 ±   0.4 ev/s (10000 events)
   985.1 ±   0.3 ev/s (10000 events)
 --------------------
   982.6 ±   5.1 ev/s
```

### This PR, with the default configuration
```
Running 4 times over 10200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
   866.1 ±   0.3 ev/s (10000 events)
   865.2 ±   0.3 ev/s (10000 events)
   859.3 ±   0.3 ev/s (10000 events)
   867.7 ±   0.3 ev/s (10000 events)
 --------------------
   864.6 ±   3.7 ev/s
```

Simply adding "jumping doublets" does not make any differences.


### This PR, with triplets
```
Running 4 times over 10200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
   525.3 ±   0.2 ev/s (10000 events)
   524.9 ±   0.2 ev/s (10000 events)
   524.6 ±   0.2 ev/s (10000 events)
   523.1 ±   0.2 ev/s (10000 events)
 --------------------
   524.5 ±   1.0 ev/s
```

### This PR, with triplets and "jumping doublets"
```
Running 4 times over 10200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
   495.8 ±   0.2 ev/s (10000 events)
   494.5 ±   0.2 ev/s (10000 events)
   494.4 ±   0.2 ev/s (10000 events)
   495.2 ±   0.2 ev/s (10000 events)
 --------------------
   495.0 ±   0.6 ev/s
```

### This PR, with triplets and iterations
```
Running 4 times over 10200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
   406.2 ±   0.2 ev/s (10000 events)
   404.9 ±   0.1 ev/s (10000 events)
   405.1 ±   0.2 ev/s (10000 events)
   406.5 ±   0.2 ev/s (10000 events)
 --------------------
   405.7 ±   0.8 ev/s
```

### This PR, with triplets, iterations and "jumping doublets"
```
Running 4 times over 10200 events with 1 jobs, each with 8 threads, 8 streams and 1 GPUs
   380.2 ±   0.2 ev/s (10000 events)
   381.3 ±   0.2 ev/s (10000 events)
   381.1 ±   0.1 ev/s (10000 events)
   379.1 ±   0.2 ev/s (10000 events)
 --------------------
   380.4 ±   1.0 ev/s
```