@fwyzard 2019-08-02
Before merging this PR, I think we should understand why the throughput without triplets drops from `982.6 ± 5.1 ev/s` to `864.6 ± 3.7 ev/s`.