@VinInn 2019-08-02
@fwyzard quadruplets can be slower than before because more code (including some more kernels) is executed and more data is accessed (notably a new flag to account if doublets/cell is connected/used). In quadruplets mode we can bisect and identify piece by piece...