@VinInn 2019-08-02
about ```iterations```:
I needed to test it as iterations are the magic great solution of CMS for tracking.
They are NOT supposed to change physics performance, only to make the wf faster.
In this case, with full doublets already built, iterations make the wf slower (as I predicted).
For what I am concerned the whole infrastructure could be removed (but somebody will ask again: what about iterations?)