@fwyzard 2019-08-02
OK, thanks.

So the two configurations we want to consider are
  - only quadruplets (as before, but throughput to be understood)
  - triplets (possibly with forward triplets/jumping pairs) without iterations