@VinInn 2019-08-02
I think that ```forward triplets/jumping pairs``` will be mandatory to satisfy POG/PAG requests to be resilient to forward (and BPIX1) inefficiencies (see single muon 1GeV realistic...)