@VinInn 2019-08-17
after removing iteration infrastructure: NO regression for Triplets

throughput Quadruplets:
reference 846
before the "fix"  758
after the "fix": 795
using lateFishbone as in reference: 760
removing ```theUsed``` 763

considering theUsed+ new earlyFishbone a speedup
this PR still misses 846-763 ev/second

one could try to add just theUsed + new earlyFishbone to reference (so no triplets)
to see what is the speed up there